feat: cancel daily-ci persistence (#138)
