+++ action-server/actions/action_daily_ci_early_opt_out.py
@@ -4,6 +4,8 @@ from rasa_sdk import Action, Tracker
 from rasa_sdk.events import SlotSet
 from rasa_sdk.executor import CollectingDispatcher
 
+from actions.lib.persistence import cancel_reminder
+
 ACTION_NAME = "action_daily_ci_early_opt_out"
 
 CANCEL_CI_SLOT = "cancel_ci"
@@ -28,6 +30,6 @@ class ActionDailyCiEarlyOptOut(Action):
             template="utter_daily_ci__early_opt_out__cancel_ci_recommendation"
         )
 
-        # TODO: cancel check-in
+        cancel_reminder(tracker.current_slot_values())
 
         return [SlotSet(CANCEL_CI_SLOT, True)]
