+++ action-server/actions/daily_ci_feel_better_form.py
@@ -6,7 +6,7 @@ from rasa_sdk.executor import CollectingDispatcher
 from rasa_sdk.forms import FormAction
 
 from actions.form_helper import request_next_slot
-from actions.lib.assessment_persistence import store_assessment
+from actions.lib.persistence import store_assessment
 
 FORM_NAME = "daily_ci_feel_better_form"
